+++ packages/ricos-common/web/src/RicosEngine.tsx
@@ -1,5 +1,5 @@
 import React, { Component, Children, FunctionComponent } from 'react';
-import themeStrategy from './themeStrategy/themeStrategy';
+import createThemeStrategy, { ThemeStrategyFunction } from './themeStrategy/themeStrategy';
 import pluginsStrategy from './pluginsStrategy/pluginsStrategy';
 import localeStrategy from './localeStrategy/localeStrategy';
 import { merge } from 'lodash';
@@ -23,9 +23,11 @@ interface EngineState {
 }
 
 export class RicosEngine extends Component<EngineProps, EngineState> {
+  themeStrategy: ThemeStrategyFunction;
   constructor(props: EngineProps) {
     super(props);
     this.state = { localeStrategy: { locale: props.locale } };
+    this.themeStrategy = createThemeStrategy();
   }
 
   static defaultProps = { locale: 'en', isMobile: false };
@@ -55,12 +57,12 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
       .map(plugin => plugin.theme)
       .filter(isDefined);
 
-    const { theme: themeStrategyResult, rawCss } = themeStrategy(
+    const { theme: themeStrategyResult, rawCss } = this.themeStrategy({
       isViewer,
       themeGeneratorFunctions,
-      theme?.palette,
-      cssOverride
-    );
+      palette: theme?.palette,
+      cssOverride,
+    });
 
     return {
       strategyProps: merge(