+++ packages/ricos-editor/web/src/RicosEditor.spec.tsx
@@ -116,7 +116,13 @@ describe('RicosEditor', () => {
     };
     const rceProps = getRCE(props).props();
     const rcePropsWrapped = getRCE(props, true).props();
-    expect(JSON.stringify(rceProps)).toStrictEqual(JSON.stringify(rcePropsWrapped));
+    // hashed theme classnames can be different; assert keys only.
+    const themeKeys = Object.keys(rceProps.theme);
+    const themeKeys_wrapped = Object.keys(rcePropsWrapped.theme);
+    expect(themeKeys).toStrictEqual(themeKeys_wrapped);
+    const rceProps_noTheme = JSON.stringify({ ...rceProps, theme: {} });
+    const rcePropsWrapped_noTheme = JSON.stringify({ ...rcePropsWrapped, theme: {} });
+    expect(rceProps_noTheme).toStrictEqual(rcePropsWrapped_noTheme);
   });
   describe('Modal API', () => {
     it('should pass openModal & closeModal to helpers', () => {